Inspc and Ctrl now member of src_emC again, full reflection generation capability here. make/mklib
* The Inspc sources were moved to an extra git repository on on 18-10-15 22:34,
  The last version with this sources is 9b173110e1c009555de194c3bda2099346426bea from 18-10-15 10:16
  Use this version to follow the revision history.
  Git has not a good handling to re-add the sources.
* The Inspc sources are moved to emC/Inspc, emC/Inspc/J2c etc, check the file names. It are consistent in comparison to 2018 and before.
* emC/Ctrl are well-tried sources, but new in this repository. There is some more, later.
* ConcurrentRingBuffer now removed (old, never used and tested)
* make/scriptsCommon with Reflection generation now contains full reflection generation. But it uses yet libs/vishiaBase.jar from the test project
* InspcTargetProxy as make/... as Visual Studio project, tested. TODO compile with gcc automatically in test.
* make/mklib for Visual Studio here, removed from Test_emC. TODO make with script, and with gcc

+++ emC/Jc/ReflectionLoadExtJc.c
@@ -18,15 +18,15 @@ Data_ExtReflectionJc* correctContent_ExtReflectionJc(ExtReflectionJc_s* ythis, M
   int baseHead;
   void const* addrBaseClasses;
   bool bOk = true;
-  int32 offsClassArray, offsReflData;
+  intptr_t offsClassArray, offsReflData;
 	STACKTRC_TENTRY("correctContent_ExtReflectionJc");
   
   /**relocate all relative addresses to absolute addresses:*/
   /**first the reference of arrayClasses. */
   baseHead = (int)(intptr_t)(extReflectionData);
-  offsClassArray = *(int32*)(&extReflectionData->arrayClasses);
+  offsClassArray = *(intptr_t*)(&extReflectionData->arrayClasses);
 	extReflectionData->arrayClasses = (ClassJc_YP const*)(offsClassArray + baseHead);
-	offsReflData = *(int*)(&extReflectionData->classDataBlock);
+	offsReflData = *(intptr_t*)(&extReflectionData->classDataBlock);
   extReflectionData->classDataBlock = (ClassJc const*)(offsReflData + baseHead); 
   if(!inRange_MemAreaC(extReflectionData->arrayClasses, minAddr, maxAddr)) {
     bOk = false;
